Rename some analyzers.
- Rename NoSpaceBeforeSemicolon analyzer to SpaceBeforeSemicolon.
- Rename SpaceAfterSemicolon analyzer to NoSpaceAfterSemicolon.
