+++ StyleChecker/StyleChecker.Test/Spacing/NoSpaceAfterSemicolon/AnalyzerTest.cs
@@ -1,23 +1,23 @@
-namespace StyleChecker.Test.Spacing.NoSpaceBeforeSemicolon
+namespace StyleChecker.Test.Spacing.NoSpaceAfterSemicolon
 {
     using System.IO;
     using Microsoft.CodeAnalysis.CodeFixes;
     using Microsoft.CodeAnalysis.Diagnostics;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using StyleChecker.Spacing.NoSpaceBeforeSemicolon;
+    using StyleChecker.Spacing.NoSpaceAfterSemicolon;
     using StyleChecker.Test.Framework;
 
     [TestClass]
     public sealed class AnalyzerTest : CodeFixVerifier
     {
-        protected override CodeFixProvider CodeFixProvider
-            => new CodeFixer();
-
         protected override DiagnosticAnalyzer DiagnosticAnalyzer
             => new Analyzer();
 
+        protected override CodeFixProvider CodeFixProvider
+            => new CodeFixer();
+
         protected override string BaseDir
-            => Path.Combine("Spacing", "NoSpaceBeforeSemicolon");
+            => Path.Combine("Spacing", "NoSpaceAfterSemicolon");
 
         [TestMethod]
         public void Empty()
@@ -30,7 +30,7 @@ namespace StyleChecker.Test.Spacing.NoSpaceBeforeSemicolon
             var fix = ReadText("CodeFix");
             Result Expected(Belief b) => b.ToResult(
                 Analyzer.DiagnosticId,
-                m => $"A white space is not needed before '{m}'");
+                m => $"A white space is needed after '{m}'");
 
             VerifyDiagnosticAndFix(code, Atmosphere.Default, Expected, fix);
         }
