+++ StyleChecker/StyleChecker.Test/Spacing/NoSpaceAfterSemicolon/Code.cs
