+++ StyleChecker/StyleChecker.Test/Spacing/NoSpaceAfterSemicolon/CodeFix.cs
