+++ StyleChecker/StyleChecker.Test/Spacing/SpaceBeforeSemicolon/AnalyzerTest.cs
@@ -1,23 +1,23 @@
-namespace StyleChecker.Test.Spacing.SpaceAfterSemicolon
+namespace StyleChecker.Test.Spacing.SpaceBeforeSemicolon
 {
     using System.IO;
     using Microsoft.CodeAnalysis.CodeFixes;
     using Microsoft.CodeAnalysis.Diagnostics;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using StyleChecker.Spacing.SpaceAfterSemicolon;
+    using StyleChecker.Spacing.SpaceBeforeSemicolon;
     using StyleChecker.Test.Framework;
 
     [TestClass]
     public sealed class AnalyzerTest : CodeFixVerifier
     {
-        protected override DiagnosticAnalyzer DiagnosticAnalyzer
-            => new Analyzer();
-
         protected override CodeFixProvider CodeFixProvider
             => new CodeFixer();
 
+        protected override DiagnosticAnalyzer DiagnosticAnalyzer
+            => new Analyzer();
+
         protected override string BaseDir
-            => Path.Combine("Spacing", "SpaceAfterSemicolon");
+            => Path.Combine("Spacing", "SpaceBeforeSemicolon");
 
         [TestMethod]
         public void Empty()
@@ -30,7 +30,7 @@ namespace StyleChecker.Test.Spacing.SpaceAfterSemicolon
             var fix = ReadText("CodeFix");
             Result Expected(Belief b) => b.ToResult(
                 Analyzer.DiagnosticId,
-                m => $"A white space is needed after '{m}'");
+                m => $"A white space is not needed before '{m}'");
 
             VerifyDiagnosticAndFix(code, Atmosphere.Default, Expected, fix);
         }
