+++ StyleChecker/StyleChecker.Test/Spacing/SpaceBeforeSemicolon/Code.cs
