+++ StyleChecker/StyleChecker.Test/Spacing/SpaceBeforeSemicolon/CodeFix.cs
