+++ StyleChecker/StyleChecker/Spacing/NoSpaceAfterSemicolon/Resources.resx
