+++ StyleChecker/StyleChecker/Spacing/SpaceBeforeSemicolon/Resources.resx
